Configurando classe `Position` para figurar como tabela do banco de dados, relacionando tabelas `Position` e `ScheduleType`, ajustando o MER e criando carga inicial para tabela `Position`

+++ api/src/main/java/apisemaperreio/escalante/repository/ScheduleTypeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import apisemaperreio.escalante.model.ScheduleType;
 
 @Repository
-public interface ScheduleTypeRepository extends JpaRepository<ScheduleType, Long> {
+public interface ScheduleTypeRepository extends JpaRepository<ScheduleType, Integer> {
 
     
 }
